refactor: form/models.rs を概念別のサブモジュールへ分割
form/models.rs に 4 つの独立した関心事 (Form 本体 / Form 設定 /
回答ポリシー / ラベル) が詰まっていて、Form 集約のコアである
ActiveForm/ArchivedForm が見つけにくくなっていた。

既存の `QuestionSet` (Form が所有する値オブジェクトだが
form/question.rs に置かれている) と同じ方針で、それぞれの値オブジェクト
が「束ねる対象」のモジュールに配置する形へ整理した:

- form/settings.rs (新規) … FormSettings / Visibility / WebhookUrl
- form/label.rs    (新規) … FormLabel / FormLabelName / FormLabelAssignment /
                              FormLabelId と認可ガード
- form/answer.rs   (追記) … AnswerSettings / AnswerVisibility /
                              AnswerAcceptancePeriod / DefaultAnswerTitle
                              (判定対象が AnswerEntry のため)
- form/models.rs   (縮小) … ActiveForm / ArchivedForm / FormMeta /
                              FormTitle / FormDescription / FormId と
                              認可ガード・Allowed impl のみ

`form/models.rs` からは上記の型を `pub use` で再エクスポートしているため、
外部からの参照パス (`form::models::FormSettings` など) は変わらない。

複数モジュールで共通利用する `is_administrator` ヘルパは form.rs に
`pub(super) fn` として移し、各サブモジュールが `use super::is_administrator`
で参照できるようにした。
