+++ server/domain/src/form/models.rs
@@ -1,36 +1,35 @@
 use chrono::{DateTime, Utc};
 #[cfg(test)]
-use common::test_utils::{arbitrary_date_time, arbitrary_opt_date_time};
+use common::test_utils::arbitrary_date_time;
 use derive_getters::Getters;
 use deriving_via::DerivingVia;
 use domain_derive::UnsafeFromRawParts;
 use errors::domain::DomainError;
 #[cfg(test)]
-use proptest::{collection::SizeRange, prelude::*, strategy::BoxedStrategy};
-#[cfg(test)]
 use proptest_derive::Arbitrary;
-use regex::Regex;
-use serde::{Deserialize, Deserializer, Serialize, de};
-use strum_macros::{Display, EnumString};
+use serde::{Deserialize, Serialize};
 use types::non_empty_string::NonEmptyString;
 
-pub use crate::form::question::{Question, QuestionSet};
+pub use crate::form::{
+    answer::{AnswerAcceptancePeriod, AnswerSettings, AnswerVisibility, DefaultAnswerTitle},
+    label::{FormLabel, FormLabelAssignment, FormLabelId, FormLabelName},
+    question::{Question, QuestionSet},
+    settings::{FormSettings, Visibility, WebhookUrl},
+};
 
 use crate::{
-    form::answer::{AnswerAuthor, AnswerEntry, AnswerTitle, PostedAnswerContents},
+    form::{
+        answer::{AnswerAuthor, AnswerEntry, AnswerTitle, PostedAnswerContents},
+        is_administrator,
+    },
     types::authorization_guard::{
         Allowed, AuthorizationGuardDefinitions, AuthorizationRole, Create, Read, SelfGuarded,
         Update,
     },
-    user::models::{Actor, Role::Administrator, User, UserId},
+    user::models::{Actor, UserId},
 };
 
-fn is_administrator(actor: &Actor) -> bool {
-    matches!(actor, Actor::User(User::ActiveUser(user)) if user.role() == &Administrator)
-}
-
 pub type FormId = types::Id<ActiveForm>;
-pub type FormLabelId = types::Id<FormLabel>;
 
 #[cfg_attr(test, derive(Arbitrary))]
 #[derive(Clone, DerivingVia, Debug, PartialOrd, PartialEq)]
@@ -57,240 +56,6 @@ impl FormDescription {
 
 #[cfg_attr(test, derive(Arbitrary))]
 #[derive(UnsafeFromRawParts, Serialize, Deserialize, Clone, Default, Debug, PartialEq)]
-pub struct FormSettings {
-    #[serde(default)]
-    webhook_url: WebhookUrl,
-    #[serde(default)]
-    visibility: Visibility,
-}
-
-impl FormSettings {
-    pub fn new() -> Self {
-        Self {
-            webhook_url: WebhookUrl::try_new(None).unwrap(),
-            visibility: Visibility::PUBLIC,
-        }
-    }
-
-    pub fn webhook_url(&self, actor: &Actor) -> Result<&WebhookUrl, DomainError> {
-        if is_administrator(actor) {
-            Ok(&self.webhook_url)
-        } else {
-            Err(DomainError::Forbidden)
-        }
-    }
-
-    pub fn visibility(&self) -> &Visibility {
-        &self.visibility
-    }
-
-    pub fn change_webhook_url(self, webhook_url: WebhookUrl) -> Self {
-        Self {
-            webhook_url,
-            ..self
-        }
-    }
-
-    pub fn change_visibility(self, visibility: Visibility) -> Self {
-        Self { visibility, ..self }
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(Clone, DerivingVia, Default, Debug, PartialEq)]
-#[deriving(From, Into, IntoInner, Serialize(via: Option::<NonEmptyString>), Deserialize(via: Option::<NonEmptyString>
-))]
-pub struct WebhookUrl(Option<NonEmptyString>);
-
-impl WebhookUrl {
-    pub fn try_new(url: Option<NonEmptyString>) -> Result<Self, DomainError> {
-        if let Some(url) = &url {
-            let regex = Regex::new("https://discord.com/api/webhooks/.*").unwrap();
-            if !regex.is_match(url) {
-                return Err(DomainError::InvalidWebhookUrl);
-            }
-        }
-
-        Ok(Self(url))
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(
-    Serialize, Deserialize, Debug, EnumString, Display, Copy, Clone, Default, PartialOrd, PartialEq,
-)]
-pub enum Visibility {
-    PUBLIC,
-    #[default]
-    PRIVATE,
-}
-
-impl TryFrom<String> for Visibility {
-    type Error = DomainError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        use std::str::FromStr;
-        Self::from_str(&value).map_err(Into::into)
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(Clone, DerivingVia, Default, Debug, PartialEq)]
-#[deriving(From, Into, IntoInner, Serialize(via: Option::<NonEmptyString>), Deserialize(via: Option::<NonEmptyString>
-))]
-pub struct DefaultAnswerTitle(Option<NonEmptyString>);
-
-impl DefaultAnswerTitle {
-    pub fn new(title: Option<NonEmptyString>) -> Self {
-        Self(title)
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(
-    Serialize, Deserialize, Debug, EnumString, Display, Copy, Clone, Default, PartialOrd, PartialEq,
-)]
-pub enum AnswerVisibility {
-    PUBLIC,
-    #[default]
-    PRIVATE,
-}
-
-impl TryFrom<String> for AnswerVisibility {
-    type Error = DomainError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        use std::str::FromStr;
-        Self::from_str(&value).map_err(Into::into)
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(Serialize, Deserialize, Getters, Clone, Default, Debug, PartialEq)]
-pub struct AnswerAcceptancePeriod {
-    #[cfg_attr(test, proptest(strategy = "arbitrary_opt_date_time()"))]
-    #[serde(default)]
-    start_at: Option<DateTime<Utc>>,
-    #[cfg_attr(test, proptest(strategy = "arbitrary_opt_date_time()"))]
-    #[serde(default)]
-    end_at: Option<DateTime<Utc>>,
-}
-
-impl AnswerAcceptancePeriod {
-    pub fn try_new(
-        start_at: Option<DateTime<Utc>>,
-        end_at: Option<DateTime<Utc>>,
-    ) -> Result<Self, DomainError> {
-        match (start_at, end_at) {
-            (Some(start_at), Some(end_at)) if start_at > end_at => {
-                Err(DomainError::InvalidAnswerAcceptancePeriod)
-            }
-            _ => Ok(Self { start_at, end_at }),
-        }
-    }
-
-    pub fn is_within_period(&self, now: DateTime<Utc>) -> bool {
-        if let Some(start_at) = self.start_at
-            && start_at > now
-        {
-            return false;
-        }
-        if let Some(end_at) = self.end_at
-            && end_at < now
-        {
-            return false;
-        }
-        true
-    }
-}
-
-/// フォームの回答にまつわる設定をまとめた値オブジェクトです。
-///
-/// 回答の公開範囲・受付期間・仮回答可否・デフォルトタイトルといった「ポリシー」を保持し、
-/// [`AnswerEntry`] の閲覧可否 ([`Self::can_read_entry`]) や新規受理 ([`Self::try_accept_answer`])
-/// を判断します。この値オブジェクトは [`ActiveForm`] が所有します。
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(Serialize, Deserialize, Getters, Clone, Default, Debug, PartialEq)]
-pub struct AnswerSettings {
-    default_answer_title: DefaultAnswerTitle,
-    visibility: AnswerVisibility,
-    acceptance_period: AnswerAcceptancePeriod,
-    allow_temporary_answers: bool,
-}
-
-impl AnswerSettings {
-    pub fn new(
-        default_answer_title: DefaultAnswerTitle,
-        visibility: AnswerVisibility,
-        acceptance_period: AnswerAcceptancePeriod,
-        allow_temporary_answers: bool,
-    ) -> Self {
-        Self {
-            default_answer_title,
-            visibility,
-            acceptance_period,
-            allow_temporary_answers,
-        }
-    }
-
-    pub fn change_default_answer_title(self, default_answer_title: DefaultAnswerTitle) -> Self {
-        Self {
-            default_answer_title,
-            ..self
-        }
-    }
-
-    pub fn change_visibility(self, visibility: AnswerVisibility) -> Self {
-        Self { visibility, ..self }
-    }
-
-    pub fn change_acceptance_period(self, acceptance_period: AnswerAcceptancePeriod) -> Self {
-        Self {
-            acceptance_period,
-            ..self
-        }
-    }
-
-    pub fn change_allow_temporary_answers(self, allow_temporary_answers: bool) -> Self {
-        Self {
-            allow_temporary_answers,
-            ..self
-        }
-    }
-
-    /// `author` / `actor` の組み合わせと受付期間・仮回答可否から、新しい [`AnswerEntry`] を
-    /// 受理してよいかを判断し、受理できる場合のみ [`AnswerEntry`] を生成します。
-    /// `actor` が `author` として回答を作成してよいかを、受付期間と一時回答の可否から判定する。
-    pub(crate) fn can_accept_answer(&self, author: &AnswerAuthor, actor: &Actor) -> bool {
-        let is_within_period = self.acceptance_period.is_within_period(Utc::now());
-
-        match (author, actor) {
-            (AnswerAuthor::AuthenticatedUser(user_id), Actor::User(User::ActiveUser(user))) => {
-                *user_id == *user.id() && (is_within_period || user.role() == &Administrator)
-            }
-            (AnswerAuthor::TemporaryUser(_), Actor::User(User::TemporaryUser(_))) => {
-                self.allow_temporary_answers && is_within_period
-            }
-            _ => false,
-        }
-    }
-
-    /// `actor` が `entry` を閲覧できるかどうかを、回答の公開範囲をもとに判断します。
-    pub fn can_read_entry(&self, entry: &AnswerEntry, actor: &Actor) -> bool {
-        match actor {
-            Actor::User(User::ActiveUser(user)) => {
-                entry.author().authenticated_user_id() == Some(*user.id())
-                    || self.visibility == AnswerVisibility::PUBLIC
-                    || user.role() == &Administrator
-            }
-            Actor::System => true,
-            _ => false,
-        }
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(UnsafeFromRawParts, Serialize, Deserialize, Clone, Default, Debug, PartialEq)]
 pub struct FormMeta {
     #[cfg_attr(test, proptest(strategy = "arbitrary_date_time()"))]
     #[serde(default = "chrono::Utc::now")]
@@ -306,55 +71,6 @@ impl FormMeta {
             created_at: Utc::now(),
             updated_at: Utc::now(),
         }
-    }
-}
-
-#[derive(Serialize, Clone, DerivingVia, Default, Debug, PartialEq)]
-#[deriving(IntoInner)]
-pub struct FormLabelAssignment(Vec<FormLabelId>);
-
-impl FormLabelAssignment {
-    pub fn try_new(label_ids: Vec<FormLabelId>) -> Result<Self, DomainError> {
-        if label_ids
-            .iter()
-            .enumerate()
-            .any(|(index, label_id)| label_ids[..index].contains(label_id))
-        {
-            return Err(DomainError::InvalidEntity {
-                message: "form label ids must be unique within a form".to_string(),
-            });
-        }
-
-        Ok(Self(label_ids))
-    }
-
-    pub fn empty() -> Self {
-        Self(Vec::new())
-    }
-
-    pub fn as_slice(&self) -> &[FormLabelId] {
-        &self.0
-    }
-}
-
-impl<'de> Deserialize<'de> for FormLabelAssignment {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        Vec::<FormLabelId>::deserialize(deserializer)
-            .and_then(|value| FormLabelAssignment::try_new(value).map_err(de::Error::custom))
-    }
-}
-
-#[cfg(test)]
-impl Arbitrary for FormLabelAssignment {
-    type Parameters = (SizeRange, <FormLabelId as Arbitrary>::Parameters);
-    type Strategy = BoxedStrategy<Self>;
-
-    fn arbitrary_with(args: Self::Parameters) -> Self::Strategy {
-        Vec::<FormLabelId>::arbitrary_with(args)
-            .prop_filter_map("unique form label ids", |value| {
-                FormLabelAssignment::try_new(value).ok()
-            })
-            .boxed()
     }
 }
 
@@ -546,7 +262,7 @@ impl AuthorizationRole for ActiveForm {
 impl AuthorizationGuardDefinitions for ActiveForm {
     /// [`ActiveForm`] の作成権限があるかどうかを判定します。
     ///
-    /// 作成権限は [`Administrator`] のみに与えられます。
+    /// 作成権限は [`Administrator`](crate::user::models::Role::Administrator) のみに与えられます。
     fn can_create(&self, actor: &Actor) -> bool {
         is_administrator(actor)
     }
@@ -556,16 +272,16 @@ impl AuthorizationGuardDefinitions for ActiveForm {
     /// 読み取り権限は以下のいずれかを満たす場合に与えられます。
     /// - [`Actor::System`] である場合
     /// - [`FormSettings`] の [`Visibility`] が [`Visibility::PUBLIC`] である場合
-    /// - [`Administrator`] である場合
+    /// - [`Administrator`](crate::user::models::Role::Administrator) である場合
     fn can_read(&self, actor: &Actor) -> bool {
         matches!(actor, Actor::System)
-            || self.settings.visibility == Visibility::PUBLIC
+            || self.settings.visibility() == &Visibility::PUBLIC
             || is_administrator(actor)
     }
 
     /// [`ActiveForm`] の更新権限があるかどうかを判定します。
     ///
-    /// 更新権限は [`Administrator`] のみに与えられます。
+    /// 更新権限は [`Administrator`](crate::user::models::Role::Administrator) のみに与えられます。
     fn can_update(&self, actor: &Actor) -> bool {
         is_administrator(actor)
     }
@@ -578,76 +294,10 @@ impl AuthorizationGuardDefinitions for ActiveForm {
     }
 }
 
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(Clone, DerivingVia, Debug, PartialOrd, PartialEq)]
-#[deriving(From, Into, IntoInner, Serialize(via: NonEmptyString), Deserialize(via: NonEmptyString))]
-pub struct FormLabelName(NonEmptyString);
-
-impl FormLabelName {
-    pub fn new(name: NonEmptyString) -> Self {
-        Self(name)
-    }
-}
-
-#[cfg_attr(test, derive(Arbitrary))]
-#[derive(UnsafeFromRawParts, Serialize, Deserialize, Getters, Debug, PartialEq)]
-pub struct FormLabel {
-    id: FormLabelId,
-    name: FormLabelName,
-}
-
-impl FormLabel {
-    pub fn new(name: FormLabelName) -> Self {
-        Self {
-            id: FormLabelId::new(),
-            name,
-        }
-    }
-
-    pub fn renamed(&self, name: FormLabelName) -> Self {
-        Self { id: self.id, name }
-    }
-}
-
-impl AuthorizationRole for FormLabel {
-    type Role = SelfGuarded;
-}
-
-impl AuthorizationGuardDefinitions for FormLabel {
-    /// [`FormLabel`] の作成権限があるかどうかを判定します。
-    ///
-    /// 作成権限は [`Administrator`] のみに与えられます。
-    fn can_create(&self, actor: &Actor) -> bool {
-        is_administrator(actor)
-    }
-
-    /// [`FormLabel`] の読み取り権限はすべてのユーザーに与えられます。
-    fn can_read(&self, _actor: &Actor) -> bool {
-        true
-    }
-
-    /// [`FormLabel`] の更新権限があるかどうかを判定します。
-    ///
-    /// 更新権限は [`Administrator`] のみに与えられます。
-    fn can_update(&self, actor: &Actor) -> bool {
-        is_administrator(actor)
-    }
-
-    /// [`FormLabel`] の削除権限があるかどうかを判定します。
-    ///
-    /// 削除権限は [`Administrator`] のみに与えられます。
-    fn can_delete(&self, actor: &Actor) -> bool {
-        is_administrator(actor)
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use chrono::Duration;
-
     use super::*;
     use crate::form::question::{Question, QuestionId, QuestionType};
-    use crate::user::models::Role;
     use types::non_empty_vec::NonEmptyVec;
     use uuid::Uuid;
 
@@ -669,160 +319,6 @@ mod tests {
             .unwrap(),
         )
         .unwrap()
-    }
-
-    fn answer_settings(
-        allow_temporary_answers: bool,
-        acceptance_period: AnswerAcceptancePeriod,
-    ) -> AnswerSettings {
-        AnswerSettings::new(
-            DefaultAnswerTitle::new(None),
-            AnswerVisibility::PRIVATE,
-            acceptance_period,
-            allow_temporary_answers,
-        )
-    }
-
-    fn active_user(role: Role) -> crate::user::models::ActiveUser {
-        crate::user::models::ActiveUser::new("user".to_string(), UserId::from(Uuid::new_v4()), role)
-    }
-
-    fn answer_entry(author: AnswerAuthor) -> AnswerEntry {
-        AnswerEntry::new(
-            FormId::new(),
-            author,
-            AnswerTitle::new(None),
-            PostedAnswerContents::try_new(&[], Vec::new()).unwrap(),
-        )
-    }
-
-    #[test]
-    fn temporary_answer_creation_requires_allow_flag() {
-        let settings = answer_settings(false, AnswerAcceptancePeriod::try_new(None, None).unwrap());
-        let author = AnswerAuthor::TemporaryUser(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-        let actor = Actor::from(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-
-        assert!(!settings.can_accept_answer(&author, &actor));
-    }
-
-    #[test]
-    fn temporary_answer_creation_succeeds_when_allowed_and_within_period() {
-        let settings = answer_settings(true, AnswerAcceptancePeriod::try_new(None, None).unwrap());
-        let author = AnswerAuthor::TemporaryUser(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-        let actor = Actor::from(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-
-        assert!(settings.can_accept_answer(&author, &actor));
-    }
-
-    #[test]
-    fn temporary_answer_creation_respects_acceptance_period() {
-        let settings = answer_settings(
-            true,
-            AnswerAcceptancePeriod::try_new(
-                Some(Utc::now() - Duration::days(2)),
-                Some(Utc::now() - Duration::days(1)),
-            )
-            .unwrap(),
-        );
-        let author = AnswerAuthor::TemporaryUser(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-        let actor = Actor::from(crate::user::models::TemporaryUser::new(
-            "guest".to_string(),
-            "contact".to_string(),
-        ));
-
-        assert!(!settings.can_accept_answer(&author, &actor));
-    }
-
-    #[test]
-    fn private_entry_is_readable_by_its_author() {
-        let author = active_user(Role::StandardUser);
-        let entry = answer_entry(AnswerAuthor::AuthenticatedUser(*author.id()));
-        let settings = answer_settings(false, AnswerAcceptancePeriod::try_new(None, None).unwrap());
-
-        assert!(settings.can_read_entry(&entry, &Actor::from(author)));
-    }
-
-    #[test]
-    fn private_entry_is_not_readable_by_other_standard_user() {
-        let author = active_user(Role::StandardUser);
-        let other = active_user(Role::StandardUser);
-        let entry = answer_entry(AnswerAuthor::AuthenticatedUser(*author.id()));
-        let settings = answer_settings(false, AnswerAcceptancePeriod::try_new(None, None).unwrap());
-
-        assert!(!settings.can_read_entry(&entry, &Actor::from(other)));
-    }
-
-    #[test]
-    fn private_entry_is_readable_by_administrator() {
-        let author = active_user(Role::StandardUser);
-        let administrator = active_user(Role::Administrator);
-        let entry = answer_entry(AnswerAuthor::AuthenticatedUser(*author.id()));
-        let settings = answer_settings(false, AnswerAcceptancePeriod::try_new(None, None).unwrap());
-
-        assert!(settings.can_read_entry(&entry, &Actor::from(administrator)));
-    }
-
-    #[test]
-    fn public_entry_is_readable_by_other_standard_user() {
-        let author = active_user(Role::StandardUser);
-        let other = active_user(Role::StandardUser);
-        let entry = answer_entry(AnswerAuthor::AuthenticatedUser(*author.id()));
-        let settings = AnswerSettings::new(
-            DefaultAnswerTitle::new(None),
-            AnswerVisibility::PUBLIC,
-            AnswerAcceptancePeriod::try_new(None, None).unwrap(),
-            false,
-        );
-
-        assert!(settings.can_read_entry(&entry, &Actor::from(other)));
-    }
-
-    #[test]
-    fn form_label_id_set_allows_empty_ids() {
-        let label_ids = FormLabelAssignment::try_new(vec![]).unwrap();
-
-        assert!(label_ids.as_slice().is_empty());
-    }
-
-    #[test]
-    fn form_label_id_set_allows_unique_ids() {
-        let first = FormLabelId::new();
-        let second = FormLabelId::new();
-        let label_ids = FormLabelAssignment::try_new(vec![first, second]).unwrap();
-
-        assert_eq!(label_ids.as_slice(), &[first, second]);
-    }
-
-    #[test]
-    fn form_label_id_set_rejects_duplicate_ids() {
-        let label_id = FormLabelId::new();
-        let result = FormLabelAssignment::try_new(vec![label_id, label_id]);
-
-        assert!(matches!(result, Err(DomainError::InvalidEntity { .. })));
-    }
-
-    #[test]
-    fn form_label_id_set_deserialize_rejects_duplicate_ids() {
-        let label_id = FormLabelId::new();
-        let serialized = serde_json::to_string(&vec![label_id, label_id]).unwrap();
-        let result = serde_json::from_str::<FormLabelAssignment>(&serialized);
-
-        assert!(result.is_err());
     }
 
     #[test]
